diff to upcast int8/int16 to float64 to be consistent with other metods

+++ pandas/tests/test_algos.py
@@ -2441,7 +2441,7 @@ class TestDiff:
     def test_diff_low_precision_int(self, dtype):
         arr = np.array([0, 1, 1, 0, 0], dtype=dtype)
         result = algos.diff(arr, 1)
-        expected = np.array([np.nan, 1, 0, -1, 0], dtype="float32")
+        expected = np.array([np.nan, 1, 0, -1, 0], dtype="float64")
         tm.assert_numpy_array_equal(result, expected)
